Some more minor fixes
