+++ build.gradle
@@ -44,6 +44,7 @@ project(':ambry-api') {
     dependencies {
         compile project(':ambry-utils')
         compile "com.codahale.metrics:metrics-core:$metricsVersion"
+        testCompile project(':ambry-utils').sourceSets.test.output
     }
 }
